pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pwick"
version = "2.1.0"
description = "A simple, secure, and 100% local password manager"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    {name = "orpheus497"}
]
keywords = ["password", "manager", "security", "encryption", "local-first", "privacy"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Security",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "PySide6",
    "cryptography>=41.0.7",
    "argon2-cffi>=23.1.0",
    "pyperclip>=1.8.2",
    "zxcvbn>=4.4.28",
    "tomli>=2.0.0; python_version < '3.11'",
    "tomli-w>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/orpheus497/pwick"
"Bug Reports" = "https://github.com/orpheus497/pwick/issues"
"Source" = "https://github.com/orpheus497/pwick"

[project.scripts]
pwick = "pwick.__main__:main"

[project.gui-scripts]
pwick-gui = "pwick.__main__:main"

[project.optional-dependencies]
test = [
    "pytest",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
